requirements.txt
# Ranger 2 + TPL 100 Check-in App
# Python 3.8+

flask>=2.3.0
requests>=2.28.0
pyyaml>=6.0
pywin32>=306; sys_platform == "win32"
